check if synonyms exists

+++ app/Http/Requests/UpdateTaxon.php
@@ -249,9 +249,14 @@ class UpdateTaxon extends FormRequest
 
     private function createSynonyms(Taxon $taxon)
     {
+        if ($this->input('synonyms') == null or $this->input('removed_synonyms') != []) {
+            return;
+        }
+
         foreach ($this->input('removed_synonyms') as $removed) {
             Synonym::find($removed['id'])->delete();
         }
+
         foreach ($this->input('synonyms') as $synonym) {
             if (isset($synonym['id'])) {
                 continue;
